requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
astroid==2.9.3
    # via pylint
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk.cloud-assembly-schema==1.144.0
    # via
    #   aws-cdk.core
    #   aws-cdk.cx-api
aws-cdk.core==1.144.0
    # via monocdk-nag
aws-cdk.cx-api==1.144.0
    # via aws-cdk.core
aws-cdk.region-info==1.144.0
    # via aws-cdk.core
bandit==1.7.2
    # via -r requirements-dev.in
black==22.1.0
    # via -r requirements-dev.in
cattrs==1.10.0
    # via jsii
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   safety
constructs==3.3.213
    # via
    #   aws-cdk.core
    #   monocdk
    #   monocdk-nag
coverage==6.3.1
    # via -r requirements-dev.in
dparse==0.5.1
    # via safety
flake8==4.0.1
    # via -r requirements-dev.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.26
    # via bandit
idna==3.3
    # via requests
importlib-metadata==4.2.0
    # via
    #   click
    #   flake8
    #   stevedore
isort==5.10.1
    # via
    #   -r requirements-dev.in
    #   pylint
jsii==1.53.0
    # via
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
    #   monocdk
    #   monocdk-nag
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
monocdk==1.144.0
    # via monocdk-nag
monocdk-nag==1.4.49
    # via -r requirements-dev.in
mypy==0.931
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   dparse
    #   safety
pathspec==0.9.0
    # via black
pbr==5.8.1
    # via stevedore
platformdirs==2.5.0
    # via
    #   black
    #   pylint
publication==0.0.3
    # via
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
    #   monocdk
    #   monocdk-nag
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pylint==2.12.2
    # via -r requirements-dev.in
pyparsing==3.0.7
    # via packaging
python-dateutil==2.8.2
    # via jsii
pyyaml==6.0
    # via
    #   bandit
    #   dparse
requests==2.27.1
    # via safety
safety==1.10.3
    # via -r requirements-dev.in
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
stevedore==3.5.0
    # via bandit
toml==0.10.2
    # via
    #   dparse
    #   pylint
tomli==2.0.1
    # via
    #   black
    #   mypy
typed-ast==1.5.2
    # via
    #   astroid
    #   black
    #   mypy
types-requests==2.27.9
    # via -r requirements-dev.in
types-urllib3==1.26.9
    # via types-requests
typing-extensions==4.0.1
    # via
    #   astroid
    #   black
    #   cattrs
    #   gitpython
    #   importlib-metadata
    #   jsii
    #   mypy
    #   pylint
urllib3==1.26.8
    # via requests
wrapt==1.13.3
    # via astroid
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
